Fix structure/naming of omega Tas/Addtask
