+++ PWGDQ/dielectron/LMEE/AliAnalysisTaskOmegaDielectron_AccEff.cxx
@@ -1,125 +1,81 @@
 #include "AliAnalysisTaskOmegaDielectron_AccEff.h"
+#include <iostream>
 
 /// \cond CLASSIMP
 ClassImp(AliAnalysisTaskOmegaDielectron_AccEff);
 /// \endcond
 
 /***************************************************************************/ /**
-                                                                               * ROOT I/O Constructor.
-                                                                               ******************************************************************************/
+* ROOT I/O Constructor.
+******************************************************************************/
 AliAnalysisTaskOmegaDielectron_AccEff::AliAnalysisTaskOmegaDielectron_AccEff()
-    : AliAnalysisTaskSE(),
-      // General member variables
-      fOutputList(nullptr), fEvent(nullptr), fMCEvent(nullptr), fEventCuts(),
-      fESDtrackCuts(nullptr), fCutMode(100),
-      // Toggles
-      fIsESD(kTRUE), fIsMC(kFALSE), fUseCent(kFALSE),
-      // Cut Parameters
-      fTriggerMask(AliVEvent::kMB | AliVEvent::kINT7), fMinEta(-0.8),
-      fMaxEta(0.8), fMinPt(0.2), fMaxPt(10.0), fMaxZv(10.0), fMinCent(0.0),
-      fMaxCent(100.0),
-      // Arrays for Binning
-      fBinsMult(nullptr), fBinsCent(nullptr), fBinsPt(nullptr), fBinsy(nullptr),
-      fBinsEta(nullptr), fBinsZv(nullptr), fBinsPtReso(nullptr),
-      //pdg codes:
-      felectron_pdg(11),fpositron_pdg(-11),
-      fmother_pdg(223),
-      // fmother_pdg(333),
-      //storage Vectors
-      v_elec_true_omega(0x0),
-      v_posi_true_omega(0x0),
-      v_elec_motherID_true_omega(0),
-      v_posi_motherID_true_omega(0),
-      // Event-Histograms
-      fHistEventSelection(nullptr),
-      fHistMC_Omegas_Rapidity(nullptr),
-      fHistMC_ele1_posi2_OmegaDielDeacay(nullptr),
-      fHistMC_Omegas_gen(nullptr),
-      fHistMC_Omegas_gen_DaughtersinAcc(nullptr),
-      fHist_rec_true_Ele_Omegas_Mothers(nullptr),
-      fHist_rec_true_Pos_Omegas_Mothers(nullptr),
-      fHist_rec_true_Dielec(nullptr)
+: AliAnalysisTaskSE(),
+// General member variables
+fOutputList(nullptr), fEvent(nullptr), fMCEvent(nullptr), fEventCuts(nullptr),
+fFilter_TrackCuts(), fFilter_PID(),
+fPIDResponse(nullptr),
+fMinEta(-0.8), fMaxEta(0.8), fMinPt(0.2), fMaxPt(10.0),
+//pdg codes:
+felectron_pdg(11),fpositron_pdg(-11),
+fmother_pdg(223),
+// fmother_pdg(333),
+//storage Vectors
+v_elec_true_omega(0x0),
+v_posi_true_omega(0x0),
+v_elec_true_omega_MCPart(0x0),
+v_posi_true_omega_MCPart(0x0),
+v_elec_motherID_true_omega(0),
+v_posi_motherID_true_omega(0),
+// Event-Histograms
+fHistEventSelection(nullptr),
+fHist_MC_Omegas_Rapidity(nullptr),
+fHist_MC_Omegas_gen(nullptr),
+fHist_MC_Omegas_gen_DaughtersinAcc(nullptr),
+fHist_MC_Omegas_TrackCuts(nullptr),
+fHist_MC_Omegas_TrackPID(nullptr),
+fHist_Rec_Omegas_TrackCuts(nullptr),
+fHist_Rec_Omegas_TrackPID(nullptr)
 {
   // ROOT IO constructor, don't allocate memory here!
 }
 
 /***************************************************************************/ /**
-                                                                               * Constructor.
-                                                                               ******************************************************************************/
+* Constructor.
+******************************************************************************/
 AliAnalysisTaskOmegaDielectron_AccEff::AliAnalysisTaskOmegaDielectron_AccEff(const char *name)
-    : AliAnalysisTaskSE(name),
-      // General member variables
-      fOutputList(nullptr), fEvent(nullptr), fMCEvent(nullptr), fEventCuts(),
-      fESDtrackCuts(nullptr), fCutMode(100),
-      // Toggles
-      fIsESD(kTRUE),
-      // fIsMC(kTRUE),
-      fIsMC(kFALSE), fUseCent(kFALSE),
-      // Cut Parameters
-      fTriggerMask(AliVEvent::kMB | AliVEvent::kINT7), fMinEta(-0.8),
-      fMaxEta(0.8), fMinPt(0.2), fMaxPt(10.0), fMaxZv(10.0), fMinCent(0.0),
-      fMaxCent(100.0),
-      // Arrays for Binning
-      fBinsMult(nullptr), fBinsCent(nullptr), fBinsPt(nullptr), fBinsy(nullptr),
-      fBinsEta(nullptr), fBinsZv(nullptr), fBinsPtReso(nullptr),
-      //pdg codes:
-      felectron_pdg(11),fpositron_pdg(-11),
-      fmother_pdg(223),
-      // fmother_pdg(333),
-      //storage Vectors
-      v_elec_true_omega(0x0),
-      v_posi_true_omega(0x0),
-      v_elec_motherID_true_omega(0),
-      v_posi_motherID_true_omega(0),
-      // Event-Histograms
-      fHistEventSelection(nullptr),
-      fHistMC_Omegas_Rapidity(nullptr),
-      fHistMC_ele1_posi2_OmegaDielDeacay(nullptr),
-      fHistMC_Omegas_gen(nullptr),
-      fHistMC_Omegas_gen_DaughtersinAcc(nullptr),
-      fHist_rec_true_Ele_Omegas_Mothers(nullptr),
-      fHist_rec_true_Pos_Omegas_Mothers(nullptr),
-      fHist_rec_true_Dielec(nullptr)
+: AliAnalysisTaskSE(name),
+// General member variables
+fOutputList(nullptr), fEvent(nullptr), fMCEvent(nullptr), fEventCuts(nullptr),
+fFilter_TrackCuts(), fFilter_PID(),
+fPIDResponse(nullptr),
+fMinEta(-0.8), fMaxEta(0.8), fMinPt(0.2), fMaxPt(10.0),
+//pdg codes:
+felectron_pdg(11),fpositron_pdg(-11),
+fmother_pdg(223),
+// fmother_pdg(333),
+//storage Vectors
+v_elec_true_omega(0x0),
+v_posi_true_omega(0x0),
+v_elec_true_omega_MCPart(0x0),
+v_posi_true_omega_MCPart(0x0),
+v_elec_motherID_true_omega(0),
+v_posi_motherID_true_omega(0),
+// Event-Histograms
+fHistEventSelection(nullptr),
+fHist_MC_Omegas_Rapidity(nullptr),
+fHist_MC_Omegas_gen(nullptr),
+fHist_MC_Omegas_gen_DaughtersinAcc(nullptr),
+fHist_MC_Omegas_TrackCuts(nullptr),
+fHist_MC_Omegas_TrackPID(nullptr),
+fHist_Rec_Omegas_TrackCuts(nullptr),
+fHist_Rec_Omegas_TrackPID(nullptr)
 {
-  // Set default binning
-  Double_t binsMultDefault[2] = {0., 10000.};
-  Double_t binsCentDefault[2] = {0., 100.};
-  Double_t binsPtDefault[53] = {
-      0.1,  0.15, 0.2,  0.25, 0.3,  0.35, 0.4,  0.45, 0.5, 0.55, 0.6,
-      0.65, 0.7,  0.75, 0.8,  0.85, 0.9,  0.95, 1.0,  1.1, 1.2,  1.3,
-      1.4,  1.5,  1.6,  1.7,  1.8,  1.9,  2.0,  2.2,  2.4, 2.6,  2.8,
-      3.0,  3.2,  3.4,  3.6,  3.8,  4.0,  4.5,  5.0,  5.5, 6.0,  6.5,
-      7.0,  8.0,  9.0,  10.0, 20.0, 30.0, 40.0, 50.0, 60.0};
-  Double_t binsEtaDefault[19] = {-0.9, -0.8, -0.7, -0.6, -0.5, -0.4, -0.3,
-                                 -0.2, -0.1, 0.,   0.1,  0.2,  0.3,  0.4,
-                                 0.5,  0.6,  0.7,  0.8,  0.9};
-  Double_t binsyDefault[19] = {-0.9, -0.8, -0.7, -0.6, -0.5, -0.4, -0.3,
-                                -0.2, -0.1, 0.,   0.1,  0.2,  0.3,  0.4,
-                                0.5,  0.6,  0.7,  0.8,  0.9};
-
-  Double_t binsZvDefault[13] = {-30., -25., -20., -15., -10., -5., 0.,
-                                5.,   10.,  15.,  20.,  25.,  30.};
-
-
-  // binning for relative pT resolution
-  const Int_t nBinsPtReso = 300;
-  Double_t binsPtReso[nBinsPtReso + 1];
-  SetFixedBinEdges(binsPtReso, 0., 0.3, nBinsPtReso);
-  SetBinsPtReso(nBinsPtReso, binsPtReso);
-
-  SetBinsMult(1, binsMultDefault);
-  SetBinsCent(1, binsCentDefault);
-  SetBinsPt(52, binsPtDefault);
-  SetBinsEta(18, binsEtaDefault);
-  SetBinsy(18, binsyDefault);
-  SetBinsZv(12, binsZvDefault);
-
   DefineOutput(1, TList::Class());
 }
 
-/***************************************************************************/ /**
-                                                                               * Function executed once before the event loop. Create histograms here.
-                                                                               ******************************************************************************/
+/***************************************************************************/
+/*** Function executed once before the event loop. Create histograms here.**
+****************************************************************************/
 void AliAnalysisTaskOmegaDielectron_AccEff::UserCreateOutputObjects() {
 
   OpenFile(1, "recreate");
@@ -127,70 +83,57 @@ void AliAnalysisTaskOmegaDielectron_AccEff::UserCreateOutputObjects() {
   fOutputList->SetOwner();
 
   // Control histogram to check the effect of event cuts
-  fHistEventSelection = new TH1F("fHistEventSelection", "fHistEventSelection [all : selected]", 2, 0.5, 2.5);
+  fHistEventSelection = new TH1F("fHistEventSelection", "fHistEventSelection [all : selected]", 2, -0.5, 1.5);
   fHistEventSelection->GetYaxis()->SetTitle("#it{N}_{events}");
-  fHistEventSelection->GetXaxis()->SetBinLabel(1, "all");
-  fHistEventSelection->GetXaxis()->SetBinLabel(2, "selected");
+  fHistEventSelection->GetXaxis()->SetBinLabel(1, "in");
+  fHistEventSelection->GetXaxis()->SetBinLabel(2, "out");
   fOutputList->Add(fHistEventSelection);
 
-  fHistMC_Omegas_Rapidity = new TH1F("fHistMC_Omegas_Rapidity", "Omega - Rapidity", 200, -10, 10);
-  fHistMC_Omegas_Rapidity->GetYaxis()->SetTitle("#it{N}_{events}");
-  fHistMC_Omegas_Rapidity->GetXaxis()->SetTitle("y");
-  fOutputList->Add(fHistMC_Omegas_Rapidity);
-
-  // if(fIsMC)
-  // {
-  fHistMC_ele1_posi2_OmegaDielDeacay = new TH1F("fHistMC_ele1_posi2_OmegaDielDeacy", "Omega - Dielectron Deacay", 2, 0.5, 2.5);
-  fHistMC_ele1_posi2_OmegaDielDeacay->GetYaxis()->SetTitle("N");
-  fHistMC_ele1_posi2_OmegaDielDeacay->GetXaxis()->SetBinLabel(1, "electrons");
-  fHistMC_ele1_posi2_OmegaDielDeacay->GetXaxis()->SetBinLabel(2, "positrons");
-  fOutputList->Add(fHistMC_ele1_posi2_OmegaDielDeacay);
-  // }
-
-  fHistMC_Omegas_gen = new TH2F("fHistMC_Omegas_gen", "fHistMC_Omegas_gen;#it{p}_{T} (GeV/#it{c});y", 50, 0, 10, 50, -10, 10);
-  fOutputList->Add(fHistMC_Omegas_gen);
-  fHistMC_Omegas_gen_DaughtersinAcc = new TH2F("fHistMC_Omegas_gen_DaughtersinAcc", "fHistMC_Omegas_gen_DaughtersinAcc;#it{p}_{T} (GeV/#it{c});y", 50, 0, 10, 50, -10, 10);
-  fOutputList->Add(fHistMC_Omegas_gen_DaughtersinAcc);
-
-  fHist_rec_true_Ele_Omegas_Mothers = new TH1F("Reconstructed_True_Electrons_Omegas_Mothers_Pt", "Reconstructed_True_Electrons_Omegas_Mothers_Pt;#it{p}_{T} (GeV/#it{c})", 50, 0, 10);
-  fHistMC_Omegas_Rapidity->GetYaxis()->SetTitle("#it{N}_{events}");
-  fOutputList->Add(fHist_rec_true_Ele_Omegas_Mothers);
-  fHist_rec_true_Pos_Omegas_Mothers = new TH1F("Reconstructed_True_Positrons_Omegas_Mothers_Pt", "Reconstructed_True_Positrons_Omegas_Mothers_Pt;#it{p}_{T} (GeV/#it{c})", 50, 0, 10);
-  fHistMC_Omegas_Rapidity->GetYaxis()->SetTitle("#it{N}_{events}");
-  fOutputList->Add(fHist_rec_true_Pos_Omegas_Mothers);
-  fHist_rec_true_Dielec = new TH2D("Reconstructed_True_Omegas_DielectronD_InvMass_Pt", "Reconstructed_True_Omegas_DielectronD_InvMass_Pt;Inv_Mass;#it{p}_{T,ee} (GeV/#it{c})", 50, 0, 10, 50, 0, 10);
-  fHistMC_Omegas_Rapidity->GetYaxis()->SetTitle("#it{N}_{events}");
-  fOutputList->Add(fHist_rec_true_Dielec);
+  // Histogram to check Rapidity distribution
+  fHist_MC_Omegas_Rapidity = new TH1F("fHist_MC_Omegas_Rapidity", "Omega - Rapidity;y;#it{N}_{events}", 200, -10, 10);
+  fOutputList->Add(fHist_MC_Omegas_Rapidity);
 
 
+  /// --- --- --- MC histograms --- --- --- ///
+  fHist_MC_Omegas_gen = new TH3D("fHist_MC_Omegas_gen", "fHist_MC_Omegas_gen;#it{m}_{ee} (GeV/#it{c}^{2});#it{p}_{T} (GeV/#it{c});y",  100, 0, 10, 100, 0, 10, 20, -1, 1);
+  fOutputList->Add(fHist_MC_Omegas_gen);
+  fHist_MC_Omegas_gen_DaughtersinAcc = new TH3D("fHist_MC_Omegas_gen_DaughtersinAcc", "fHist_MC_Omegas_gen_DaughtersinAcc;#it{m}_{ee} (GeV/#it{c}^{2});#it{p}_{T} (GeV/#it{c});y", 100, 0, 10, 100, 0, 10, 20, -1, 1);
+  fOutputList->Add(fHist_MC_Omegas_gen_DaughtersinAcc);
 
-  // override event automatic event selection settings
+  // m - pt - y
+  fHist_MC_Omegas_TrackCuts = new TH3D("fHist_MC_Omegas_TrackCuts", "fHist_MC_Omegas_TrackCuts;#it{m}_{ee} (GeV/#it{c}^{2});#it{p}_{T,ee} (GeV/#it{c});y", 100, 0, 10, 100, 0, 10, 20, -1, 1);
+  fOutputList->Add(fHist_MC_Omegas_TrackCuts);
+  fHist_MC_Omegas_TrackPID = new TH3D("fHist_MC_Omegas_TrackPID", "fHist_MC_Omegas_TrackPID;#it{m}_{ee} (GeV/#it{c}^{2});#it{p}_{T,ee} (GeV/#it{c});y",  100, 0, 10, 100, 0, 10, 20, -1, 1);
+  fOutputList->Add(fHist_MC_Omegas_TrackPID);
+
+  /// --- --- --- histograms -> reconstructed Omega --- --- --- ///
+  // m- pt-y
+  fHist_Rec_Omegas_TrackCuts = new TH3D("fHist_Rec_Omegas_TrackCuts", "fHist_Rec_Omegas_TrackCuts;#it{m}_{ee} (GeV/#it{c}^{2});#it{p}_{T,ee} (GeV/#it{c});y", 100, 0, 10, 100, 0, 10, 20, -1, 1);
+  fOutputList->Add(fHist_Rec_Omegas_TrackCuts);
+  fHist_Rec_Omegas_TrackPID = new TH3D("fHist_Rec_Omegas_TrackPID", "fHist_Rec_Omegas_TrackPID;#it{m}_{ee} (GeV/#it{c}^{2});#it{p}_{T,ee} (GeV/#it{c});y", 100, 0, 10, 100, 0, 10, 20, -1, 1);
+  fOutputList->Add(fHist_Rec_Omegas_TrackPID);
 
 
-  ///??????????????///, 60, -0.5, 6 - 0.5);
-  fEventCuts.SetMaxVertexZposition(fMaxZv);
-  if (fUseCent)
-    fEventCuts.SetCentralityRange(fMinCent, fMaxCent);
-  fEventCuts.OverrideAutomaticTriggerSelection(fTriggerMask);
+  // Set EventCuts -> needed for :: AliDielectronPID
+  fEventCuts=new AliDielectronEventCuts("eventCuts","Vertex Track && |vtxZ|<10 && ncontrib>0");
+  fEventCuts->SetVertexType(AliDielectronEventCuts::kVtxSPD); // AOD
+  fEventCuts->SetRequireVertex();
+  fEventCuts->SetVertexZ(-10.,10.);
+  fEventCuts->SetMinVtxContributors(1);
 
-    ///??????????????///
-  if (fIsESD)
-    InitESDTrackCuts();
+  InitCuts();
 
   PostData(1, fOutputList);
 }
 
 /// Destructor
 AliAnalysisTaskOmegaDielectron_AccEff::~AliAnalysisTaskOmegaDielectron_AccEff() {
-  if (fESDtrackCuts) {
-    delete fESDtrackCuts;
-    fESDtrackCuts = nullptr;
-  }
+
 }
 
 /***************************************************************************/ /**
-                                                                               * Function executed for each event.
-                                                                               ******************************************************************************/
+* Function executed for each event.
+******************************************************************************/
 void AliAnalysisTaskOmegaDielectron_AccEff::UserExec(Option_t *) {
 
   fEvent = InputEvent();
@@ -199,54 +142,52 @@ void AliAnalysisTaskOmegaDielectron_AccEff::UserExec(Option_t *) {
     return;
   }
 
-  // if(fIsMC){
-  // AliMCEvent* fMCEvent = MCEvent();
+  AliInputEventHandler *eventHandler = nullptr;
+  eventHandler = dynamic_cast<AliAODInputHandler*> (AliAnalysisManager::GetAnalysisManager()->GetInputEventHandler());
+
+  if (!fPIDResponse) SetPIDResponse( eventHandler->GetPIDResponse() );
+  AliDielectronVarManager::SetPIDResponse(fPIDResponse);
+
   fMCEvent = MCEvent();
+  AliDielectronVarManager::SetEvent(fMCEvent);
   if (!fMCEvent) {
     AliError("fMCEvent not available\n");
     return;
   }
-  // }
+
+
 
 
   ///-------------------- Fill Events in Histogramm :: All;Selected;???? --------------------///
 
   fHistEventSelection->Fill(1.0); // all events
-  if (!fEventCuts.AcceptEvent(fEvent)) // https://github.com/alisw/AliPhysics/blob/master/OADB/AliEventCuts.cxx
-    return;
+  // if (!fEventCuts->IsSelected(fMCEvent))
+  if (!fEventCuts->IsSelected(fEvent)) // https://github.com/alisw/AliPhysics/blob/master/OADB/AliEventCuts.cxx
+  return;
   fHistEventSelection->Fill(2.0); // selected events
 
 
   ///-------------------- Loop over Generated MC Particles ---OMEGAS --------------------///
-  // if (fIsMC){
-  double parity = 0.8;
 
   for(Int_t iGenPart = 0; iGenPart < fMCEvent->GetNumberOfTracks();iGenPart++) {
     AliMCParticle* mcGenParticle  = (AliMCParticle*)fMCEvent->GetTrack(iGenPart);
     if(!mcGenParticle){
-      AliError("mcGenParticle  not available\n");
+      AliError("mcGenParticle not available\n");
       continue;
     }
 
     Int_t pdgcode_gen = mcGenParticle->PdgCode();
     Int_t mother_id_gen = ( mcGenParticle->GetLabel() );
 
-    if(pdgcode_gen == fmother_pdg){
-      fHistMC_Omegas_Rapidity->Fill(mcGenParticle->Y());
-    }
+    if(pdgcode_gen == fmother_pdg){  // Check if MC Particle is an omega
+      fHist_MC_Omegas_Rapidity->Fill(mcGenParticle->Y());
 
-
-    if ( TMath::Abs( mcGenParticle->Y() ) >= parity ) continue;
-
-    if(pdgcode_gen == fmother_pdg){
-      if( CheckDielectronDecay(mcGenParticle) ){
-        fHistMC_Omegas_gen->Fill(mcGenParticle->Pt(), mcGenParticle->Y());
-        // std::cout << "mother_id Gen: " << mother_id_gen << '\n';
+      if( CheckDielectronDecay(mcGenParticle) ){ // Check if :: w->e+e-
+        fHist_MC_Omegas_gen->Fill(mcGenParticle->M(), mcGenParticle->Pt(), mcGenParticle->Y());
       }
 
-      if( CheckDielectronDecay_DaughterinAcc(mcGenParticle) ){
-        fHistMC_Omegas_gen_DaughtersinAcc->Fill(mcGenParticle->Pt(), mcGenParticle->Y());
-        // std::cout << "mother_id Gen with daughters within /eta/<0.8 : " << mother_id_gen << '\n';
+      if( CheckDielectronDecay_DaughterinAcc(mcGenParticle) ){ // Check if :: w->e+e- with /n_e/<0.8 & 0.2<p_T,e < 10GeV/c
+        fHist_MC_Omegas_gen_DaughtersinAcc->Fill(mcGenParticle->M(), mcGenParticle->Pt(), mcGenParticle->Y());
       }
 
       else{
@@ -258,30 +199,31 @@ void AliAnalysisTaskOmegaDielectron_AccEff::UserExec(Option_t *) {
 
 
 
-  ///-------------------- Loop over AOD Tracks --------------------///
+
+  ///-------------------- Loop over AOD Tracks - electrons// positrons --------------------///
 
   AliAODTrack *track = nullptr;
-  // Create a vector containing the electron Track
+  // Create a vector containing the electron/positron tracks
   v_elec_true_omega.clear();
+  v_elec_true_omega_MCPart.clear();
   v_elec_motherID_true_omega.clear();
   v_posi_true_omega.clear();
+  v_posi_true_omega_MCPart.clear();
   v_posi_motherID_true_omega.clear();
 
   for (Int_t iTrack = 0; iTrack < fEvent->GetNumberOfTracks(); iTrack++) {
     track = static_cast<AliAODTrack*>(fEvent->GetTrack(iTrack));
-    if (!track || !track->TestFilterBit(128)){  // filter bit 128 denotes TPC-only tracks, use only them for the analysis
-      // AliErrorF("Could not receive track %d\n", iTrack);
-      continue;
-    }
-    track = static_cast<AliAODTrack*>(track);
-    if (!AcceptKinematics(track)){
+
+    if (!track || !track->TestFilterBit(16)){ // SetAODFilterBit(AliDielectronTrackCuts::kGlobalNoDCA); // 1<<4 // -> FilterBit 16
+      // filter bit 128 denotes TPC-only tracks, use only them for the analysis
       continue;
     }
 
-
+    if (!AcceptKinematics(track)){ // Check if electron/positron  /n_e/<0.8 & 0.2<p_T,e < 10GeV/c
+      continue;
+    }
 
     ///--- Find original particle in MC-Stack ---///
-    // if(fIsMC){
     Int_t mcLabel =  TMath::Abs(track->GetLabel()); // negative label means bad quality track
     AliMCParticle *mcParticle = (AliMCParticle *)fMCEvent->GetTrack(mcLabel);
     if (!mcParticle) {
@@ -289,7 +231,6 @@ void AliAnalysisTaskOmegaDielectron_AccEff::UserExec(Option_t *) {
       continue;
     }
     if ( !mcParticle->IsPhysicalPrimary() ){
-      // AliError("mcParticle not physical primary\n");
       continue;
     }
 
@@ -298,27 +239,24 @@ void AliAnalysisTaskOmegaDielectron_AccEff::UserExec(Option_t *) {
     Int_t mother_id = ( mcParticle_mother->GetLabel() );
     Int_t pdgCode_mother = ( mcParticle_mother->PdgCode() );
 
-////***************ELECTRONS*****************////
-    if(pdgCode_mother == fmother_pdg){
-      // std::cout << "Mother is Omega" << '\n';
-
-      if(pdgcode == felectron_pdg){
-        if(CheckDielectronDecay(mcParticle_mother)){
-          fHistMC_ele1_posi2_OmegaDielDeacay->Fill(1.0);
+    ////***************ELECTRONS*****************////
+    if(pdgCode_mother == fmother_pdg){                        // Check if mother of MC Particle (Track) is an omega
+      if(pdgcode == felectron_pdg){                           // Check if MC Particle (Track) is an electron
+        if(CheckDielectronDecay(mcParticle_mother)){          // Check if omega has a 2-body dielectron decay
           v_elec_true_omega.push_back(track);
+          v_elec_true_omega_MCPart.push_back(mcParticle_mother);
           v_elec_motherID_true_omega.push_back(mother_id);
-          std::cout << "--------::  electron track: " << iTrack << "  electron label: " << mcLabel << "  electron mother: " << mother_id << '\n';
         }
       }
-////***************POSITRONS*****************////
+      ////***************POSITRONS*****************////
       if(pdgcode == fpositron_pdg){
         if(CheckDielectronDecay(mcParticle_mother)){
-          fHistMC_ele1_posi2_OmegaDielDeacay->Fill(2.0);
           v_posi_true_omega.push_back(track);
+          v_posi_true_omega_MCPart.push_back(mcParticle_mother);
           v_posi_motherID_true_omega.push_back(mother_id);
-          std::cout << "--------::  positron track: " << iTrack << "  positron label: " << mcLabel << "  positron mother: " << mother_id << '\n';
         }
       }
+
       else{
         continue;
       }
@@ -326,43 +264,29 @@ void AliAnalysisTaskOmegaDielectron_AccEff::UserExec(Option_t *) {
 
   }
 
-  // //Check vectors
-  // if (v_elec_true_omega.size() >0){
-  //   for( auto const & item : v_elec_true_omega)
-  //       std::cout<<item<<", BLAAAAAAAAAAAAAAAAA"<< '\n';
-  // }
-  // if (v_posi_true_omega.size() >0){
-  //   for( auto const & item_pos : v_posi_true_omega)
-  //       std::cout<<item_pos<<", BLAAAAAAAAAAAAAAAAAMMMMMMMMMMM"<< '\n';
-  // }
 
+
+  ///-------------------- Pair electron + positron of the same Omegamother  --------------------///
+  ///--------------------  Go over the selected electron/positron Tracks    --------------------///
 
   if (v_posi_true_omega.size() == 0) return;
   if (v_elec_true_omega.size() == 0) return;
   else{
+    /// ~~~~~~~ electron vectors ~~~~~~~ ///
     for(int i=0; i< (Int_t)v_elec_true_omega.size(); i++){
-    // for(auto const & i : v_elec_true_omega){
       AliAODTrack * ele_from_same_mother_track= (AliAODTrack *)v_elec_true_omega.at(i);
+      AliMCParticle *mcParticle_same_mother = (AliMCParticle *)v_elec_true_omega_MCPart.at(i); // Omega mother!
       int mother_electron= (Int_t)v_elec_motherID_true_omega.at(i);
 
+      /// ~~~~~~~ positron vectors ~~~~~~~ ///
       for(int j=0; j< (Int_t)v_posi_true_omega.size(); j++){
-      // for(auto const & j : v_posi_true_omega){
         AliAODTrack * pos_from_same_mother_track= (AliAODTrack *)v_posi_true_omega.at(j);
+        // AliMCParticle *mcParticle_p_same_mother = (AliMCParticle *)v_posi_true_omega_MCPart.at(i); // Omega mother!
         int mother_positron= (Int_t)v_posi_motherID_true_omega.at(j);
 
         if(mother_electron==mother_positron){
-          fHist_rec_true_Ele_Omegas_Mothers->Fill(ele_from_same_mother_track->Pt());
-          fHist_rec_true_Pos_Omegas_Mothers->Fill(pos_from_same_mother_track->Pt());
-
-          AliMCParticle *mcParticle_ele_trail = (AliMCParticle *)fMCEvent->GetTrack(TMath::Abs(ele_from_same_mother_track->GetLabel()));
-          std::cout << " Electron pdg: " << mcParticle_ele_trail->PdgCode() << "  mclabel: " << TMath::Abs(ele_from_same_mother_track->GetLabel()) << '\n';
-          AliMCParticle *mcParticle_pos_trail = (AliMCParticle *)fMCEvent->GetTrack(TMath::Abs(pos_from_same_mother_track->GetLabel()));
-          std::cout << " Positron pdg: " << mcParticle_pos_trail->PdgCode() << "  mclabel: " << TMath::Abs(pos_from_same_mother_track->GetLabel()) << '\n';
-
-          // Double_t electron_mass=0.00051099895;
           Double_t electron_mass = (AliPID::ParticleMass(AliPID::kElectron) );
-          // std::cout << "ELECTRON  pt: " << Pt_elec << "  eta: " << Eta_elec << "  phi: " << Phi_elec << " m: " << electron_mass << '\n';
-          // std::cout << "POSITRON  pt: " << Pt_posi << "  eta: " << Eta_posi << "  phi: " << Phi_posi << " m: " << electron_mass << '\n';
+          //Tracks:
           TLorentzVector Lvec1;
           Lvec1.SetPtEtaPhiM(ele_from_same_mother_track->Pt(), ele_from_same_mother_track->Eta(), ele_from_same_mother_track->Phi(), electron_mass);
           TLorentzVector Lvec2;
@@ -371,10 +295,32 @@ void AliAnalysisTaskOmegaDielectron_AccEff::UserExec(Option_t *) {
           TLorentzVector LvecM = Lvec1 + Lvec2;
           Double_t mass = LvecM.M();
           Double_t pairpt = LvecM.Pt();
+          Double_t pairy = LvecM.Rapidity();
           Double_t weight = 1;
           std::cout << "  mass: " << mass << "  pairpt: " << pairpt << '\n';
 
-          fHist_rec_true_Dielec->Fill(mass, pairpt);
+          /// ~~~~~~ ***** pass Track cuts *****  ~~~~~~ ///
+          UInt_t selectedMask_Track =( 1 << fFilter_TrackCuts->GetCuts()->GetEntries())-1; // cutting logic taken from AliDielectron::FillTrackArrays()   // apply track cuts
+          UInt_t cutMask_electron = fFilter_TrackCuts->IsSelected(ele_from_same_mother_track);
+          UInt_t cutMask_positron = fFilter_TrackCuts->IsSelected(pos_from_same_mother_track);
+          if (cutMask_electron == selectedMask_Track && cutMask_positron == selectedMask_Track) {
+            fHist_Rec_Omegas_TrackCuts->Fill(mass, pairpt, pairy);
+            // omega histo
+            cout << mcParticle_same_mother->M()<< '\n';
+            fHist_MC_Omegas_TrackCuts->Fill(mcParticle_same_mother->M(), mcParticle_same_mother->Pt(), mcParticle_same_mother->Y());
+
+
+            /// ~~~~~~ ***** pass PID cuts *****  ~~~~~~ ///
+            UInt_t selectedMask_PID =( 1 << fFilter_PID->GetCuts()->GetEntries())-1;
+            UInt_t cutMask_PID_electron = fFilter_PID->IsSelected(ele_from_same_mother_track);
+            UInt_t cutMask_PID_positron = fFilter_PID->IsSelected(pos_from_same_mother_track);
+            if (cutMask_PID_electron == selectedMask_PID && cutMask_PID_positron == selectedMask_PID) {
+              fHist_Rec_Omegas_TrackPID->Fill(mass, pairpt, pairy);
+              // omega histo
+              fHist_MC_Omegas_TrackPID->Fill(mcParticle_same_mother->M(), mcParticle_same_mother->Pt(), mcParticle_same_mother->Y());
+            }
+          }
+
         }
         else{
           continue;
@@ -382,8 +328,6 @@ void AliAnalysisTaskOmegaDielectron_AccEff::UserExec(Option_t *) {
       }
     }
   }
-
-
 
 
 
@@ -396,23 +340,9 @@ void AliAnalysisTaskOmegaDielectron_AccEff::UserExec(Option_t *) {
 *********************************************************/
 void AliAnalysisTaskOmegaDielectron_AccEff::Terminate(Option_t *) {}
 
-/***************************************************************************/ /**
-* Function to select primary charged particles.
-******************************************************************************/
-Bool_t AliAnalysisTaskOmegaDielectron_AccEff::IsChargedPrimary(Int_t mcLabel) {
-  AliMCParticle *mcParticle = (AliMCParticle *)fMCEvent->GetTrack(mcLabel);
-  if (!mcParticle->IsPhysicalPrimary()){
-    return kFALSE;
-  }
-  if (!mcParticle) {
-    AliError("mcGenParticle  not available\n");
-    return kFALSE;
-  }
-  if (!(TMath::Abs(mcParticle->Charge()) > 0.01))
-    return kFALSE;
-  return kTRUE;
-}
-
+// /***************************************************************************/ /**
+// * Function to Check if the particle decays via dielectrons
+// ******************************************************************************/
 Bool_t AliAnalysisTaskOmegaDielectron_AccEff::CheckDielectronDecay(AliMCParticle *particle) {
   // Get the daughters:
   Int_t number_daughters_bool = particle->GetNDaughters();
@@ -421,6 +351,11 @@ Bool_t AliAnalysisTaskOmegaDielectron_AccEff::CheckDielectronDecay(AliMCParticle
 
   bool ele_from_same_mother_bool = false;
   bool pos_from_same_mother_bool = false;
+
+  double rapidi = 0.8;
+  if ( TMath::Abs( particle->Y() ) >= rapidi ){
+    return false;
+  } // omega :: /y/<0.8
 
   if(number_daughters_bool != 2){
     return false;
@@ -435,16 +370,19 @@ Bool_t AliAnalysisTaskOmegaDielectron_AccEff::CheckDielectronDecay(AliMCParticle
         pos_from_same_mother_bool = true;
       }
     }
-    // if (ele_from_same_mother_bool == true && pos_from_same_mother_bool == true && number_daughters_bool==2){
+
     if (ele_from_same_mother_bool && pos_from_same_mother_bool ){
       return true; // true -> if daughters ARE Dieceltrons!
       // (number_daughters==2 Wichtig! )
     }
     else
-      return false; // false -> if daughters are not e+ + e-
+    return false; // false -> if daughters are not e+ + e-
   }
 }
 
+// /***************************************************************************/ /**
+// * Function to Check if the particle decays via dielectrons, which are within the Acceptence
+// ******************************************************************************/
 Bool_t AliAnalysisTaskOmegaDielectron_AccEff::CheckDielectronDecay_DaughterinAcc(AliMCParticle *particle) {
   // Get the daughters:
   Int_t number_daughters_bool = particle->GetNDaughters();
@@ -457,6 +395,10 @@ Bool_t AliAnalysisTaskOmegaDielectron_AccEff::CheckDielectronDecay_DaughterinAcc
   bool ele_from_same_mother_bool = false;
   bool pos_from_same_mother_bool = false;
 
+  double rapidi = 0.8;
+  if ( TMath::Abs( particle->Y() ) >= rapidi ){
+    return false;
+  } // omega :: /y/<0.8
 
   if(number_daughters_bool != 2){
     return false;
@@ -473,439 +415,151 @@ Bool_t AliAnalysisTaskOmegaDielectron_AccEff::CheckDielectronDecay_DaughterinAcc
         daughter_pos = (AliMCParticle*)fMCEvent->GetTrack(daughter_i_bool);
       }
     }
+
     if (ele_from_same_mother_bool == true && pos_from_same_mother_bool == true && AcceptKinematics(daughter_elec) && AcceptKinematics(daughter_pos) ){ // eta und pt cut
-    // if (ele_from_same_mother_bool == true && pos_from_same_mother_bool == true && EtaCut(daughter_elec) && EtaCut(daughter_pos) ){ // eta cut
       return true; // true -> if daughters ARE Dieceltrons! within Acceptance
       // (number_daughters==2 Wichtig! )
     }
     else
-      return false; // false -> if daughters are not e+ + e-
+    return false; // false -> if daughters are not e+ + e-
   }
 }
 
 /***************************************************************************/ /**
-                                                                               * Function to select if the track or mc particle is in defined kinematic range.
-                                                                               ******************************************************************************/
+* Function to select if the track or mc particle is in defined kinematic range.
+******************************************************************************/
 Bool_t AliAnalysisTaskOmegaDielectron_AccEff::AcceptKinematics(AliVParticle *particle) {
   if (!particle)
-    return kFALSE;
+  return kFALSE;
 
   Double_t eta = particle->Eta();
   Double_t pt = particle->Pt();
 
   if (eta <= fMinEta + PRECISION)
-    return kFALSE;
+  return kFALSE;
   if (eta >= fMaxEta - PRECISION)
-    return kFALSE;
+  return kFALSE;
   if (pt <= fMinPt + PRECISION)
-    return kFALSE;
+  return kFALSE;
   if (pt >= fMaxPt - PRECISION)
-    return kFALSE;
+  return kFALSE;
   return kTRUE;
 }
 
-Bool_t AliAnalysisTaskOmegaDielectron_AccEff::EtaCut(AliVParticle *particle) {
-  if (!particle)
-    return kFALSE;
+// /***************************************************************************/ /**
+// * Function to initialize the cuts.
+// ******************************************************************************/
+void AliAnalysisTaskOmegaDielectron_AccEff::InitCuts() {
+  std::cout << "InitCuts()" <<std::endl;
 
-  Double_t eta = particle->Eta();
+  // TrackCuts (in one)
+  if (fFilter_TrackCuts)
+  delete fFilter_TrackCuts;
+  fFilter_TrackCuts = new AliAnalysisFilter("Analysis Filter _ fFilter_TrackCuts");
 
-  if (eta <= fMinEta + PRECISION)
-    return kFALSE;
-  if (eta >= fMaxEta - PRECISION)
-    return kFALSE;
-  return kTRUE;
-}
-
-
-
-
-
-
-/***************************************************************************/ /**
-                                                                               * Function to select tracks with required quality.
-                                                                               ******************************************************************************/
-Bool_t AliAnalysisTaskOmegaDielectron_AccEff::AcceptTrackQuality(AliVTrack *track) {
-  if (fIsESD) {
-    AliESDtrack *esdTrack = dynamic_cast<AliESDtrack *>(track);
-    if (!fESDtrackCuts->AcceptTrack(esdTrack))
-      return kFALSE;
-  }
-  return kTRUE;
-}
-
-/***************************************************************************/ /**
-                                                                               * Function to obtain V0M centrality.
-                                                                               ******************************************************************************/
-Double_t AliAnalysisTaskOmegaDielectron_AccEff::GetCentrality(AliVEvent *event) {
-  Double_t centrality = -1;
-  AliMultSelection *multSelection =
-      (AliMultSelection *)fEvent->FindListObject("MultSelection");
-  if (!multSelection) {
-    AliError("No MultSelection found!");
-    return 999;
-  }
-  centrality = multSelection->GetMultiplicityPercentile("V0M");
-  if (centrality > 100) {
-    AliError("Centrality determination does not work proprely!");
-    return 999;
-  }
-  return centrality;
-}
-
-/***************************************************************************/ /**
-                                                                               * Function to initialize the ESD track cuts object.
-                                                                               ******************************************************************************/
-void AliAnalysisTaskOmegaDielectron_AccEff::InitESDTrackCuts() {
-
-  if (fESDtrackCuts)
-    delete fESDtrackCuts;
-  fESDtrackCuts = new AliESDtrackCuts("AliESDtrackCuts");
-  if (!fESDtrackCuts) {
-    AliError("fESDtrackCuts not available");
+  if (!fFilter_TrackCuts) {
+    AliError("fFilter_TrackCuts not available");
     return;
   }
+  fFilter_TrackCuts->AddCuts(SetupTrackCuts());
 
-  fESDtrackCuts->SetRequireTPCRefit(kTRUE);
-  fESDtrackCuts->SetMinRatioCrossedRowsOverFindableClustersTPC(0.8);
-  fESDtrackCuts->SetMaxChi2PerClusterTPC(4.0);
-  fESDtrackCuts->SetMaxFractionSharedTPCClusters(0.4);
-  fESDtrackCuts->SetRequireITSRefit(kTRUE);
-  fESDtrackCuts->SetClusterRequirementITS(AliESDtrackCuts::kSPD,
-                                          AliESDtrackCuts::kAny);
-  fESDtrackCuts->SetMaxChi2PerClusterITS(36.);
-  fESDtrackCuts->SetDCAToVertex2D(kFALSE);
-  fESDtrackCuts->SetRequireSigmaToVertex(kFALSE);
-  fESDtrackCuts->SetMaxDCAToVertexZ(2.0);
-  fESDtrackCuts->SetMaxDCAToVertexXYPtDep(
-      "0.0182+0.0350/pt^1.01"); // 7*(0.0026+0.0050/pt^1.01)
-  fESDtrackCuts->SetAcceptKinkDaughters(kFALSE);
-  fESDtrackCuts->SetMaxChi2TPCConstrainedGlobal(36.); // tpcc cut
-  fESDtrackCuts->SetCutGeoNcrNcl(3, 130, 1.5, 0.85,
-                                 0.7); // Geometrical-Length Cut
 
-  // cut-variations:
-  if (fCutMode == 101) {
-    fESDtrackCuts->SetMaxChi2PerClusterITS(25.);
-  }
-  if (fCutMode == 102) {
-    fESDtrackCuts->SetMaxChi2PerClusterITS(49.);
-  }
+  // PIDCuts
+  if (fFilter_PID)
+  delete fFilter_PID;
+  fFilter_PID = new AliAnalysisFilter("Analysis Filter _ fFilter_PID");
 
-  if (fCutMode == 103) {
-    fESDtrackCuts->SetMaxChi2PerClusterTPC(3.0);
-  }
-  if (fCutMode == 104) {
-    fESDtrackCuts->SetMaxChi2PerClusterTPC(5.0);
-  }
-
-  if (fCutMode == 105) {
-    fESDtrackCuts->SetMinRatioCrossedRowsOverFindableClustersTPC(0.7);
-  }
-  if (fCutMode == 106) {
-    fESDtrackCuts->SetMinRatioCrossedRowsOverFindableClustersTPC(0.9);
-  }
-
-  if (fCutMode == 107) {
-    fESDtrackCuts->SetMaxFractionSharedTPCClusters(0.2);
-  }
-  if (fCutMode == 108) {
-    fESDtrackCuts->SetMaxFractionSharedTPCClusters(1.0);
-  }
-
-  if (fCutMode == 109) {
-    fESDtrackCuts->SetMaxChi2TPCConstrainedGlobal(25.);
-  }
-  if (fCutMode == 110) {
-    fESDtrackCuts->SetMaxChi2TPCConstrainedGlobal(49.);
-  }
-
-  if (fCutMode == 111) {
-    fESDtrackCuts->SetMaxDCAToVertexXYPtDep("0.0104+0.0200/pt^1.01");
-  }
-  if (fCutMode == 112) {
-    fESDtrackCuts->SetMaxDCAToVertexXYPtDep("0.0260+0.0500/pt^1.01");
-  }
-
-  if (fCutMode == 113) {
-    fESDtrackCuts->SetMaxDCAToVertexZ(1.0);
-  }
-  if (fCutMode == 114) {
-    fESDtrackCuts->SetMaxDCAToVertexZ(5.0);
-  }
-
-  if (fCutMode == 115) {
-    fESDtrackCuts->SetClusterRequirementITS(AliESDtrackCuts::kSPD,
-                                            AliESDtrackCuts::kOff);
-  }
-
-  if (fCutMode == 116) {
-    fESDtrackCuts->SetCutGeoNcrNcl(3, 120, 1.5, 0.85, 0.7);
-  }
-  if (fCutMode == 117) {
-    fESDtrackCuts->SetCutGeoNcrNcl(3, 140, 1.5, 0.85, 0.7);
-  }
-
-  if (fCutMode == 118) {
-    fESDtrackCuts->SetCutGeoNcrNcl(4, 130, 1.5, 0.85, 0.7);
-  }
-  if (fCutMode == 119) {
-    fESDtrackCuts->SetCutGeoNcrNcl(2, 130, 1.5, 0.85, 0.7);
-  }
-}
-
-/***************************************************************************/ /**
-                                                                               * Function to get array of equidistant bin edges between lower and upper edge.
-                                                                               ******************************************************************************/
-void AliAnalysisTaskOmegaDielectron_AccEff::SetFixedBinEdges(Double_t *array,
-                                             Double_t lowerEdge,
-                                             Double_t upperEdge, Int_t nBins) {
-  for (Int_t i = 0; i <= nBins; i++) {
-    array[i] = lowerEdge + i * (upperEdge - lowerEdge) / nBins;
-  }
-}
-
-/***************************************************************************/ /**
-                                                                               * Function to create THnSparseF histogram with the specified axes.
-                                                                               ******************************************************************************/
-THnSparseF *AliAnalysisTaskOmegaDielectron_AccEff::CreateHistogram(string name,
-                                                   vector<string> axes) {
-  Int_t nAxes = axes.size();
-  if (nAxes > MAX_HISTO_DIM)
-    return nullptr;
-
-  Int_t nBins[MAX_HISTO_DIM] = {0};
-  Double_t lowerBounds[MAX_HISTO_DIM] = {0.0};
-  Double_t upperBounds[MAX_HISTO_DIM] = {0.0};
-
-  string title = name + " [";
-  // first figure out number of bins and dimensions
-  for (Int_t i = 0; i < nAxes; i++) {
-    TArrayD *binEdges = GetBinEdges(axes[i]);
-    nBins[i] = binEdges->GetSize() - 1;
-    lowerBounds[i] = binEdges->GetAt(0);
-    upperBounds[i] = binEdges->GetAt(binEdges->GetSize() - 1);
-    title += axes[i];
-    if (i < nAxes - 1)
-      title += " : ";
-    else
-      title += "]";
-  }
-  // create histogram
-  THnSparseF *histogram = new THnSparseF(name.c_str(), title.c_str(), nAxes,
-                                         nBins, lowerBounds, upperBounds);
-
-  // set histogram axes
-  for (Int_t i = 0; i < nAxes; i++) {
-    TArrayD *binEdges = GetBinEdges(axes[i]);
-    histogram->SetBinEdges(i, binEdges->GetArray());
-    histogram->GetAxis(i)->SetTitle(GetAxisTitle(axes[i]).c_str());
-  }
-  histogram->Sumw2();
-  return histogram;
-}
-
-/***************************************************************************/ /**
-                                                                               * Function to obtain the correct binning for the respective axis.
-                                                                               ******************************************************************************/
-TArrayD *AliAnalysisTaskOmegaDielectron_AccEff::GetBinEdges(string &axisName) {
-  if (axisName.find("sigmapt") != string::npos)
-    return fBinsPtReso;
-  else if (axisName.find("deltapt") != string::npos)
-    return fBinsPtReso;
-  else if (axisName.find("pt") != string::npos)
-    return fBinsPt;
-  else if (axisName.find("eta") != string::npos)
-    return fBinsEta;
-  else if (axisName.find("y") != string::npos)
-    return fBinsy;
-  else if (axisName.find("mult") != string::npos)
-    return fBinsMult;
-  else if (axisName.find("cent") != string::npos)
-    return fBinsCent;
-  else if (axisName.find("zv") != string::npos)
-    return fBinsZv;
-  else
-    return nullptr;
-}
-
-/***************************************************************************/ /**
-                                                                               * Function to get the correct title for each histogram axis.
-                                                                               ******************************************************************************/
-string AliAnalysisTaskOmegaDielectron_AccEff::GetAxisTitle(string &axisName) {
-  if (axisName == "pt")
-    return "#it{p}_{T} (GeV/#it{c})";
-  else if (axisName == "deltapt")
-    return "#Delta(#it{p}_{T}) / #it{p}^{ meas}_{T}";
-  else if (axisName == "mult")
-    return "Multiplicity";
-  else if (axisName == "cent")
-    return "Centrality (%)";
-  else if (axisName == "eta_meas")
-    return "#eta^{ meas}";
-  else if (axisName == "eta_true")
-    return "#eta^{ true}";
-  else if (axisName == "pt_meas")
-    return "#it{p}^{ meas}_{T} (GeV/#it{c})";
-  else if (axisName == "pt_true")
-    return "#it{p}^{ true}_{T} (GeV/#it{c})";
-  else if (axisName == "mult_meas")
-    return "#it{N}^{ meas}_{ch}";
-  else if (axisName == "mult_true")
-    return "#it{N}^{ true}_{ch}";
-  else if (axisName == "sigmapt")
-    return "#sigma(#it{p}^{ meas}_{T}) / #it{p}^{ meas}_{T}";
-  else
-    return "dummyTitle";
-}
-
-/***************************************************************************/ /**
-                                                                               * Function to fill a histogram.
-                                                                               ******************************************************************************/
-void AliAnalysisTaskOmegaDielectron_AccEff::FillHisto(THnSparseF *histo,
-                                      array<Double_t, MAX_HISTO_DIM> values) {
-  histo->Fill(values.data());
-}
-
-/***************************************************************************/ /**
-                                                                               * Function to set variable binning for multiplicity.
-                                                                               ******************************************************************************/
-void AliAnalysisTaskOmegaDielectron_AccEff::SetBinsMult(vector<Int_t> multSteps,
-                                        vector<Int_t> multBinWidth) {
-  if (multSteps.size() != multBinWidth.size()) {
-    AliError("SetBinsMult:: Vectors need to have same size!");
+  if (!fFilter_PID) {
+    AliError("fFilter_PID not available");
     return;
   }
-
-  Int_t nMultSteps = multSteps.size();
-  Int_t nBinsMult = 1; // for mult=0 bin
-  for (Int_t multBins : multSteps)
-    nBinsMult += multBins;
-  Double_t *multBinEdges = new Double_t[nBinsMult + 1]; // edges need one more
-
-  multBinEdges[0] = -0.5;
-  multBinEdges[1] = 0.5;
-  Int_t startBin = 1;
-  Int_t endBin = 1;
-  for (Int_t multStep = 0; multStep < nMultSteps; multStep++) {
-    endBin += multSteps[multStep];
-    for (Int_t multBin = startBin; multBin < endBin; multBin++) {
-      multBinEdges[multBin + 1] =
-          multBinEdges[multBin] + multBinWidth[multStep];
-    }
-    startBin = endBin;
-  }
-  SetBinsMult(nBinsMult, multBinEdges);
+  fFilter_PID->AddCuts(SetPIDcuts());
 }
 
-/***************************************************************************/ /**
-                                                                               * Function to set maxMult single multiplicity steps.
-                                                                               ******************************************************************************/
-void AliAnalysisTaskOmegaDielectron_AccEff::SetBinsMult(Int_t maxMult) {
-  return SetBinsMult({maxMult}, {1});
+// /***************************************************************************/ /**
+// * Track cuts.
+// ******************************************************************************/
+AliAnalysisCuts *AliAnalysisTaskOmegaDielectron_AccEff::SetupTrackCuts(){
+
+
+  std::cout << "SetupTrackCuts()" <<std::endl;
+  // AliAnalysisCuts* trackCuts=0x0;
+
+  AliDielectronTrackCuts *trackCutsDiel = new AliDielectronTrackCuts("trackCutsDiel","trackCutsDiel");
+  trackCutsDiel->SetAODFilterBit(AliDielectronTrackCuts::kGlobalNoDCA); // 1<<4
+  trackCutsDiel->SetClusterRequirementITS(AliDielectronTrackCuts::kSPD,AliDielectronTrackCuts::kFirst);
+  trackCutsDiel->SetRequireITSRefit(kTRUE);
+  trackCutsDiel->SetRequireTPCRefit(kTRUE);
+
+
+  AliDielectronVarCuts* trackCutsAOD =new AliDielectronVarCuts("trackCutsAOD","trackCutsAOD");
+  trackCutsAOD->AddCut(AliDielectronVarManager::kEta, -0.8,   0.8);
+  trackCutsAOD->AddCut(AliDielectronVarManager::kPt, 0.2,   10.);
+  trackCutsAOD->AddCut(AliDielectronVarManager::kImpactParXY, -1.0,   1.0); // needs in UserExec() -->  AliDielectronVarManager::SetEvent(fMCEvent);
+  trackCutsAOD->AddCut(AliDielectronVarManager::kImpactParZ,  -3.0,   3.0); // needs in UserExec() -->  AliDielectronVarManager::SetEvent(fMCEvent);
+
+  // Number of ITS clusters
+  trackCutsAOD->AddCut(AliDielectronVarManager::kNclsITS,      3.0, 999.0);
+
+  // Chi2 per ITS cluster
+  trackCutsAOD->AddCut(AliDielectronVarManager::kITSchi2Cl,    -999.,   4.5);
+
+  // Number of shared ITS clusters
+  trackCutsAOD->AddCut(AliDielectronVarManager::kNclsSITS, 1.0, 6.0, kTRUE);
+
+  // Min TPCcls
+  trackCutsAOD->AddCut(AliDielectronVarManager::kNclsTPC,      80.0, 999.0);
+
+  //Min TPC cross rows
+  trackCutsAOD->AddCut(AliDielectronVarManager::kNFclsTPCr,    100.0, 999.0);
+
+  // Min ratio TPC cross row over findable
+  trackCutsAOD->AddCut(AliDielectronVarManager::kNFclsTPCfCross,     0.8, 999.);
+
+  // Shared TPC clusters
+  trackCutsAOD->AddCut(AliDielectronVarManager::kNclsSFracTPC,     -999., 0.4);
+
+  // chi2 per TPC cls
+  trackCutsAOD->AddCut(AliDielectronVarManager::kTPCchi2Cl,    -999.,   4.0);
+
+
+  AliDielectronCutGroup* trackCuts = new AliDielectronCutGroup("Trackcuts","Trackcuts",AliDielectronCutGroup::kCompAND);
+  trackCuts->AddCut(trackCutsAOD);
+  trackCuts->AddCut(trackCutsDiel);
+
+
+  return trackCuts;
+
 }
 
-/***************************************************************************/ /**
-                                                                               * Function to add this task to a train.
-                                                                               ******************************************************************************/
-AliAnalysisTaskOmegaDielectron_AccEff *
-AliAnalysisTaskOmegaDielectron_AccEff::AddTaskMultDepSpec(TString controlstring,
-                                          Int_t cutModeLow, Int_t cutModeHigh) {
-  AliAnalysisManager *mgr = AliAnalysisManager::GetAnalysisManager();
-  if (!mgr) {
-    ::Error("AddTaskMultDepSpec", "No analysis manager found.");
-    return nullptr;
-  }
-  if (!mgr->GetInputEventHandler()) {
-    ::Error("AddTaskMultDepSpec", "No input event handler found.");
-    return nullptr;
-  }
-  TString type =
-      mgr->GetInputEventHandler()->GetDataType(); // can be "ESD" or "AOD"
-  Bool_t isMC =
-      (AliAnalysisManager::GetAnalysisManager()->GetMCtruthEventHandler() !=
-       nullptr);
 
-  // Default cut settings:
-  UInt_t triggerMask = AliVEvent::kMB | AliVEvent::kINT7;
+// /***************************************************************************/ /**
+// * PID cuts.
+// ******************************************************************************/
+AliAnalysisCuts *AliAnalysisTaskOmegaDielectron_AccEff::SetPIDcuts(){
 
-  Double_t cutVertexZ = 10.0;
+  // AliAnalysisCuts *fancyCut = 0x0;
 
-  Double_t cutCentLow = 0.0;
-  Double_t cutCentHigh = 100.0;
+  AliDielectronPID *mastermind_TPC = new AliDielectronPID("mastermind_TPC","mastermind_TPC");
+  AliDielectronPID *mastermind_TOF = new AliDielectronPID("mastermind_TOF","mastermind_TOF");
 
-  Double_t cutPtLow = 0.15;
-  Double_t cutPtHigh = 50.0;
+  //TPC electrons: includes electrons and exclude all possible other contributions using the TPC
+  mastermind_TPC->AddCut(AliDielectronPID::kTPC,AliPID::kElectron,  -3. ,3. ,0.0, 100., kFALSE,AliDielectronPID::kRequire    ,AliDielectronVarManager::kPt);
+  mastermind_TPC->AddCut(AliDielectronPID::kTPC,AliPID::kPion,    -100. ,4.,0.0, 100., kTRUE ,AliDielectronPID::kRequire    ,AliDielectronVarManager::kPt);
+  mastermind_TPC->AddCut(AliDielectronPID::kTPC,AliPID::kKaon,    -3. ,3.,0.0, 100., kTRUE, AliDielectronPID::kRequire    ,AliDielectronVarManager::kPt);
+  mastermind_TPC->AddCut(AliDielectronPID::kTPC,AliPID::kMuon,    -3. ,3.,0.0, 100., kTRUE, AliDielectronPID::kRequire    ,AliDielectronVarManager::kPt);
+  mastermind_TPC->AddCut(AliDielectronPID::kTPC,AliPID::kProton,    -3. ,3.,0.0, 100., kTRUE, AliDielectronPID::kRequire    ,AliDielectronVarManager::kPt);
 
-  Double_t cutEtaLow = -0.8;
-  Double_t cutEtaHigh = 0.8;
+  // //TOF electrons: includes all electrons, exlcludes Pions using the TPC
+  mastermind_TOF->AddCut(AliDielectronPID::kTPC,AliPID::kElectron, -3., 3. , 0. ,100., kFALSE,AliDielectronPID::kRequire    ,AliDielectronVarManager::kPt);
+  mastermind_TOF->AddCut(AliDielectronPID::kTPC,AliPID::kPion, -100, 4. , 0.0 ,100., kTRUE,AliDielectronPID::kRequire    ,AliDielectronVarManager::kPt);
+  mastermind_TOF->AddCut(AliDielectronPID::kTOF,AliPID::kElectron, -3. , 3. , 0. ,100., kFALSE,AliDielectronPID::kRequire    ,AliDielectronVarManager::kPt);
 
-  Int_t maxMult = 100;
-  string colsys = "pp";
+  // Combine
+  AliDielectronCutGroup* mastermind_cg = new AliDielectronCutGroup("mastermind_cg","mastermind_cg",AliDielectronCutGroup::kCompOR);
+  mastermind_cg->AddCut(mastermind_TPC);
+  mastermind_cg->AddCut(mastermind_TOF);
 
-  Bool_t useCent = kFALSE;
-  if (controlstring.Contains("useCent"))
-    useCent = kTRUE;
-  Double_t centBinEdges[9] = {0., 5., 10., 20., 40., 60., 80., 90., 100.};
+  return mastermind_cg;
 
-  // colison system specific settings
-  if (controlstring.Contains("pp")) {
-    colsys = "pp";
-    maxMult = 100;
-  } else if (controlstring.Contains("pPb")) {
-    colsys = "pPb";
-    maxMult = 200;
-  } else if (controlstring.Contains("XeXe")) {
-    colsys = "XeXe";
-    maxMult = 3500;
-  } else if (controlstring.Contains("PbPb")) {
-    colsys = "PbPb";
-    maxMult = 4500;
-  }
-
-  AliAnalysisTaskOmegaDielectron_AccEff *returnTask = nullptr;
-
-  char taskName[100] = "";
-
-  for (Int_t cutMode = cutModeLow; cutMode <= cutModeHigh; cutMode++) {
-    sprintf(taskName, "multDepSpec_%s_cutMode_%d", colsys.c_str(), cutMode);
-
-    AliAnalysisTaskOmegaDielectron_AccEff *task = new AliAnalysisTaskOmegaDielectron_AccEff(taskName);
-    if (cutMode == cutModeLow)
-      returnTask = task; // return one of the tasks
-
-    task->SetCutMode(cutMode);
-    task->SetTriggerMask(triggerMask);
-
-    task->SetIsMC(isMC);
-    if (type.Contains("ESD"))
-      task->SetUseESD();
-    else
-      task->SetUseAOD();
-
-    task->SetBinsMult(maxMult);
-
-    if (useCent) {
-      task->SetMinCent(cutCentLow);
-      task->SetMaxCent(cutCentHigh);
-      task->SetBinsCent(8, centBinEdges);
-    }
-
-    // kinematic cuts:
-    task->SetMinEta(cutEtaLow);
-    task->SetMaxEta(cutEtaHigh);
-    task->SetMinPt(cutPtLow);
-    task->SetMaxPt(cutPtHigh);
-    task->SetMaxZv(cutVertexZ);
-
-    // hang task in train
-    mgr->AddTask(task);
-    mgr->ConnectInput(task, 0, mgr->GetCommonInputContainer());
-    mgr->ConnectOutput(
-        task, 1,
-        mgr->CreateContainer(taskName, TList::Class(),
-                             AliAnalysisManager::kOutputContainer,
-                             "AnalysisResults.root"));
-  }
-  return returnTask;
 }
