--- PWGDQ/dielectron/macrosLMEE/AddMyTask_mwaelde_AcceptanceEfficiency.C
@@ -1,31 +1,0 @@
-AliAnalysisTask* AddMyTask_mwaelde_AcceptanceEfficiency(TString name = "name")
-{
-    // get the manager via the static access member. since it's static, you don't need
-    // to create an instance of the class here to call the function
-    AliAnalysisManager *mgr = AliAnalysisManager::GetAnalysisManager();
-    if (!mgr) {
-        return 0x0;
-    }
-    // get the input event handler, again via a static method.
-    // this handler is part of the managing system and feeds events
-    // to your task
-    if (!mgr->GetInputEventHandler()) {
-        return 0x0;
-    }
-    // by default, a file is open for writing. here, we get the filename
-    TString fileName = AliAnalysisManager::GetCommonFileName();
-    fileName += ":MyTask";      // create a subfolder in the file
-    // now we create an instance of your task
-    AliAnalysisTaskOmegaDielectron_AccEff* task = new AliAnalysisTaskOmegaDielectron_AccEff(name.Data());
-    if(!task) return 0x0;
-    task->SelectCollisionCandidates(AliVEvent::kINT7);// kAnyINT
-    // add your task to the manager
-    mgr->AddTask(task);
-    // your task needs input: here we connect the manager to your task
-    mgr->ConnectInput(task,0,mgr->GetCommonInputContainer());
-    // same for the output
-    mgr->ConnectOutput(task,1,mgr->CreateContainer("MyOutputContainer", TList::Class(), AliAnalysisManager::kOutputContainer, fileName.Data()));
-    // in the end, this macro returns a pointer to your task. this will be convenient later on
-    // when you will run your analysis in an analysis train on grid
-    return task;
-}
